requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.5.0
opencv-python>=4.7.0
albumentations>=1.3.0
scikit-learn>=1.0.0
tensorboard>=2.12.0 